refactor: AnswerApiClient 책임 분리
