+++ src/main/java/mju/iphak/maru_egg/answer/application/AnswerApiClient.java
@@ -11,17 +11,12 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.reactive.function.BodyInserters;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
-import mju.iphak.maru_egg.answer.domain.Answer;
-import mju.iphak.maru_egg.answer.dto.request.CreateAnswerRequest;
 import mju.iphak.maru_egg.answer.dto.request.LLMAskQuestionRequest;
 import mju.iphak.maru_egg.answer.dto.response.LLMAnswerResponse;
-import mju.iphak.maru_egg.answer.repository.AnswerRepository;
 import mju.iphak.maru_egg.common.exception.custom.webClient.BadRequestWebClientException;
 import mju.iphak.maru_egg.common.exception.custom.webClient.InternalServerErrorWebClientException;
 import mju.iphak.maru_egg.common.exception.custom.webClient.NotFoundWebClientException;
-import mju.iphak.maru_egg.question.domain.Question;
 import reactor.core.publisher.Mono;
 
 @RequiredArgsConstructor
@@ -29,15 +24,7 @@ import reactor.core.publisher.Mono;
 @Transactional
 public class AnswerApiClient {
 
-	private final AnswerRepository answerRepository;
 	private final WebClient webClient;
-
-	@Transactional(readOnly = true)
-	public Answer getAnswerByQuestionId(Long questionId) {
-		return answerRepository.findByQuestionId(questionId)
-			.orElseThrow(() -> new EntityNotFoundException(
-				String.format(NOT_FOUND_ANSWER_BY_QUESTION_ID.getMessage(), questionId)));
-	}
 
 	public Mono<LLMAnswerResponse> askQuestion(LLMAskQuestionRequest request) {
 		MultiValueMap<String, String> formData = new LinkedMultiValueMap<>();
@@ -62,21 +49,5 @@ public class AnswerApiClient {
 				};
 			})
 			.bodyToMono(LLMAnswerResponse.class);
-	}
-
-	public void updateAnswerContent(final Long id, final String content) {
-		Answer answer = answerRepository.findById(id)
-			.orElseThrow(() -> new EntityNotFoundException(
-				String.format(NOT_FOUND_ANSWER.getMessage(), id)));
-		answer.updateContent(content);
-	}
-
-	public void createAnswer(final Question question, final CreateAnswerRequest request) {
-		Answer answer = request.toEntity(question);
-		saveAnswer(answer);
-	}
-
-	public Answer saveAnswer(Answer answer) {
-		return answerRepository.save(answer);
 	}
 }
