+++ src/main/java/mju/iphak/maru_egg/question/application/QuestionService.java
@@ -14,6 +14,7 @@ import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import mju.iphak.maru_egg.answer.application.AnswerApiClient;
+import mju.iphak.maru_egg.answer.application.AnswerManager;
 import mju.iphak.maru_egg.answer.domain.Answer;
 import mju.iphak.maru_egg.answer.dto.response.AnswerResponse;
 import mju.iphak.maru_egg.common.dto.pagination.SliceQuestionResponse;
@@ -33,11 +34,12 @@ public class QuestionService {
 
 	private final QuestionRepository questionRepository;
 	private final AnswerApiClient answerApiClient;
+	private final AnswerManager answerManager;
 
 	public List<QuestionListItemResponse> getQuestions(final QuestionType type, final QuestionCategory category) {
 		List<Question> questions = findQuestions(type, category);
 		return questions.stream()
-			.map(question -> createQuestionResponse(question, answerApiClient.getAnswerByQuestionId(question.getId())))
+			.map(question -> createQuestionResponse(question, answerManager.getAnswerByQuestionId(question.getId())))
 			.collect(Collectors.toList());
 	}
 
@@ -61,7 +63,7 @@ public class QuestionService {
 	public void createQuestion(final CreateQuestionRequest request) {
 		Question question = request.toEntity();
 		questionRepository.save(question);
-		answerApiClient.createAnswer(question, request.answer());
+		answerManager.createAnswer(question, request.answer());
 	}
 
 	private List<Question> findQuestions(final QuestionType type, final QuestionCategory category) {
