+++ src/test/java/mju/iphak/maru_egg/answer/application/AnswerApiClientTest.java
@@ -1,12 +1,10 @@
 package mju.iphak.maru_egg.answer.application;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -19,24 +17,17 @@ import org.springframework.web.reactive.function.client.ClientResponse;
 import org.springframework.web.reactive.function.client.ExchangeFunction;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import jakarta.persistence.EntityNotFoundException;
 import mju.iphak.maru_egg.answer.domain.Answer;
-import mju.iphak.maru_egg.answer.dto.request.CreateAnswerRequest;
 import mju.iphak.maru_egg.answer.dto.request.LLMAskQuestionRequest;
 import mju.iphak.maru_egg.answer.dto.response.AnswerReferenceResponse;
 import mju.iphak.maru_egg.answer.dto.response.LLMAnswerResponse;
-import mju.iphak.maru_egg.answer.repository.AnswerRepository;
 import mju.iphak.maru_egg.common.MockTest;
 import mju.iphak.maru_egg.question.domain.Question;
 import mju.iphak.maru_egg.question.domain.QuestionCategory;
 import mju.iphak.maru_egg.question.domain.QuestionType;
-import mju.iphak.maru_egg.question.dto.request.CreateQuestionRequest;
 import reactor.core.publisher.Mono;
 
 public class AnswerApiClientTest extends MockTest {
-
-	@Mock
-	private AnswerRepository answerRepository;
 
 	@Mock
 	private ExchangeFunction exchangeFunction;
@@ -62,7 +53,7 @@ public class AnswerApiClientTest extends MockTest {
 			.baseUrl(llmBaseUrl)
 			.build();
 
-		answerApiClient = new AnswerApiClient(answerRepository, webClient);
+		answerApiClient = new AnswerApiClient(webClient);
 
 		ClientResponse clientResponse = ClientResponse.create(HttpStatusCode.valueOf(200))
 			.header("Content-Type", "application/json")
@@ -72,34 +63,6 @@ public class AnswerApiClientTest extends MockTest {
 			.build();
 
 		when(exchangeFunction.exchange(any())).thenReturn(Mono.just(clientResponse));
-	}
-
-	@DisplayName("questionId로 답변을 조회합니다.")
-	@Test
-	public void 답변_조회_성공() {
-		// given
-		when(answerRepository.findByQuestionId(1L)).thenReturn(Optional.of(answer));
-
-		// when
-		Answer result = answerApiClient.getAnswerByQuestionId(1L);
-
-		// then
-		assertNotNull(result);
-		assertEquals(answer, result);
-	}
-
-	@DisplayName("questionId로 답변 조회에 실패한 경우.")
-	@Test
-	public void 답변_조회_실패_NOTFOUND() {
-		// given
-		when(answerRepository.findByQuestionId(1L)).thenReturn(Optional.empty());
-
-		// when & then
-		EntityNotFoundException exception = assertThrows(EntityNotFoundException.class, () -> {
-			answerApiClient.getAnswerByQuestionId(1L);
-		});
-
-		assertEquals("질문 id가 1인 답변을 찾을 수 없습니다.", exception.getMessage());
 	}
 
 	@DisplayName("LLM 서버에 질문을 요청합니다.")
@@ -121,40 +84,5 @@ public class AnswerApiClientTest extends MockTest {
 		assertThat(result).isNotNull();
 		assertThat(result.answer()).isEqualTo(answer.getContent());
 		assertThat(expectedResponse).isEqualTo(result);
-	}
-
-	@DisplayName("답변 내용을 수정에 성공한 경우")
-	@Test
-	public void 답변_내용_수정_성공() throws Exception {
-		// given
-		when(answerRepository.findById(1L)).thenReturn(Optional.of(answer));
-		Long id = 1L;
-
-		// when
-		String updateContent = "변경된 답변";
-		answerApiClient.updateAnswerContent(id, updateContent);
-
-		// then
-		assertThat(answer.getContent()).isEqualTo(updateContent);
-	}
-
-	@DisplayName("질문 생성에 성공한 경우")
-	@Test
-	public void 답변_생성_성공() {
-		// given
-		CreateAnswerRequest answerRequest = new CreateAnswerRequest("example answer content", 2024);
-		CreateQuestionRequest request = new CreateQuestionRequest("example content", QuestionType.SUSI,
-			QuestionCategory.ADMISSION_GUIDELINE, answerRequest);
-		Question question = request.toEntity();
-		Answer answer = answerRequest.toEntity(question);
-
-		when(answerRepository.save(answer))
-			.thenReturn(answer);
-
-		// when
-		answerApiClient.createAnswer(question, answerRequest);
-
-		// then
-		verify(answerRepository, times(1)).save(any(Answer.class));
 	}
 }
